requirements.txt
pytest
pytest-bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.6.3-cp37-cp37m-win32.whl
allure-pytest
